reorganized main.dart into multiple files and appview folder structure. adds runZoneGuarded
