+++ lib/app/view/view.dart
@@ -1,0 +1,1 @@
+export 'app_view.dart';
