+++ lib/bootstrap.dart
@@ -1,0 +1,20 @@
+import 'dart:async';
+import 'dart:developer';
+import 'package:firebase_core/firebase_core.dart';
+import 'package:flutter/material.dart';
+import 'package:guide_solve/firebase_options.dart';
+
+Future<void> bootstrap(FutureOr<Widget> Function() builder) async{
+  WidgetsFlutterBinding.ensureInitialized();
+
+  await Firebase.initializeApp(
+    options: DefaultFirebaseOptions.currentPlatform,
+  );
+  // Bloc.observer = AppBlocObserver();
+  runZonedGuarded (() async {
+
+
+    runApp(await builder());
+  }, (error, stack) => log('$error', name: 'Error', stackTrace: stack));
+  
+}
